Added my code from my individual project.
PlotterActivity now includes my methods amendments to the activity_plotter.xml.

added a few lines of code to the CameraActivity.java file, of which the error needs to be replaced by the imageview of the photograph just taken - I think this is how it needs to be done, at least for sending the imageview over to another activity.

Submit button click listener in the PlotterActivity.java method needs to link to the next activity, of which will calculate the angles, as well as submitting the data returned, along with the array storing the positions of the pointers from PlotterActivity class, as well as the imageview of the photograph.